Add common methods
