+++ pubber/src/main/java/daq/pubber/MqttDevice.java
@@ -14,9 +14,9 @@ public class MqttDevice {
   public static final String TEST_PROJECT = "test-project";
   static final String ATTACH_TOPIC = "attach";
   public static final String CONFIG_TOPIC = "config";
-  public static final String ERRORS_TOPIC = "errors";
+  static final String ERRORS_TOPIC = "errors";
   public static final String EVENTS_TOPIC = "events";
-  public static final String STATE_TOPIC = "state";
+  static final String STATE_TOPIC = "state";
 
   private final String deviceId;
   private final Publisher publisher;
@@ -31,15 +31,7 @@ public class MqttDevice {
         prefix -> publisher.setDeviceTopicPrefix(deviceId, prefix));
   }
 
-  /**
-   * Constructs a new MqttDevice with the specified device ID and copies the publisher from the
-   * given target device.
-   * The certificate manager is set to null for this device.
-   *
-   * @param deviceId the unique identifier for the device
-   * @param target   the MqttDevice to copy the publisher from
-   */
-  public MqttDevice(String deviceId, MqttDevice target) {
+  MqttDevice(String deviceId, MqttDevice target) {
     this.deviceId = deviceId;
     publisher = target.publisher;
     certManager = null;
