+++ pubber/src/main/java/daq/pubber/Pubber.java
@@ -72,7 +72,6 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Function;
 import org.apache.http.ConnectionClosedException;
@@ -121,8 +120,21 @@ public class Pubber extends ManagerBase implements PubberHost {
   static final Date DEVICE_START_TIME = PubberHost.getRoundedStartTime();
   static final int MESSAGE_REPORT_INTERVAL = 10;
   private static final String HOSTNAME = System.getenv("HOSTNAME");
+  private static final int STATE_THROTTLE_MS = 2000;
   private static final String PUBSUB_SITE = "PubSub";
-
+  private static final int DEFAULT_REPORT_SEC = 10;
+  private static final String SYSTEM_CATEGORY_FORMAT = "system.%s.%s";
+  private static final ImmutableMap<Class<?>, String> MESSAGE_TOPIC_SUFFIX_MAP =
+      new Builder<Class<?>, String>()
+          .put(State.class, STATE_TOPIC)
+          .put(ExtraSystemState.class, STATE_TOPIC) // Used for badState option
+          .put(SystemEvents.class, PubberHost.getEventsSuffix("system"))
+          .put(PointsetEvents.class, PubberHost.getEventsSuffix("pointset"))
+          .put(ExtraPointsetEvent.class, PubberHost.getEventsSuffix("pointset"))
+          .put(InjectedMessage.class, PubberHost.getEventsSuffix("racoon"))
+          .put(InjectedState.class, STATE_TOPIC)
+          .put(DiscoveryEvents.class, PubberHost.getEventsSuffix("discovery"))
+          .build();
   private static final Map<String, String> INVALID_REPLACEMENTS = ImmutableMap.of(
       "events/blobset", "\"\"",
       "events/discovery", "{}",
@@ -135,7 +147,9 @@ public class Pubber extends ManagerBase implements PubberHost {
   private static final long RESTART_DELAY_MS = 1000;
   private static final String CORRUPT_STATE_MESSAGE = "!&*@(!*&@!";
   private static final long INJECT_MESSAGE_DELAY_MS = 2000; // Delay to make sure testing is stable.
+  private static final int FORCED_STATE_TIME_MS = 10000;
   private static final Duration CLOCK_SKEW = Duration.ofMinutes(30);
+  private static final Duration SMOKE_CHECK_TIME = Duration.ofMinutes(5);
   private static final int STATE_SPAM_SEC = 5; // Expected config-state response time.
   final State deviceState = new State();
   final Config deviceConfig = new Config();
@@ -412,6 +426,54 @@ public class Pubber extends ManagerBase implements PubberHost {
     }
   }
 
+  @Override
+  public void update(Object update) {
+    requireNonNull(update, "null update message");
+    boolean markerClass = update instanceof Class<?>;
+    final Object checkValue = markerClass ? null : update;
+    final Object checkTarget;
+    try {
+      checkTarget = markerClass ? ((Class<?>) update).getConstructor().newInstance() : update;
+    } catch (Exception e) {
+      throw new RuntimeException("Could not create marker instance of class " + update.getClass());
+    }
+    if (checkTarget == this) {
+      publishSynchronousState();
+    } else if (checkTarget instanceof SystemState) {
+      deviceState.system = (SystemState) checkValue;
+      ifTrueThen(options.dupeState, this::sendDupeState);
+    } else if (checkTarget instanceof PointsetState) {
+      deviceState.pointset = (PointsetState) checkValue;
+    } else if (checkTarget instanceof LocalnetState) {
+      deviceState.localnet = (LocalnetState) checkValue;
+    } else if (checkTarget instanceof GatewayState) {
+      deviceState.gateway = (GatewayState) checkValue;
+    } else if (checkTarget instanceof DiscoveryState) {
+      deviceState.discovery = (DiscoveryState) checkValue;
+    } else {
+      throw new RuntimeException(
+          "Unrecognized update type " + checkTarget.getClass().getSimpleName());
+    }
+    markStateDirty();
+  }
+
+  private void sendDupeState() {
+    State dupeState = new State();
+    dupeState.system = deviceState.system;
+    dupeState.timestamp = deviceState.timestamp;
+    dupeState.version = deviceState.version;
+    publishStateMessage(dupeState);
+  }
+
+  @Override
+  public void publish(Object message) {
+    publishDeviceMessage(message);
+  }
+
+  public void publish(String targetId, Object message) {
+    publishDeviceMessage(targetId, message);
+  }
+
   private void pullDeviceMessage() {
     while (true) {
       try {
@@ -564,6 +626,14 @@ public class Pubber extends ManagerBase implements PubberHost {
     }
   }
 
+  private void captureExceptions(String action, Runnable runnable) {
+    try {
+      runnable.run();
+    } catch (Exception e) {
+      error(action, e);
+    }
+  }
+
   @Override
   public void startConnection(Function<String, Boolean> connectionDone) {
     String nonce = String.valueOf(System.currentTimeMillis());
@@ -641,6 +711,14 @@ public class Pubber extends ManagerBase implements PubberHost {
     captureExceptions("disconnecting mqtt", this::disconnectMqtt);
   }
 
+  private void disconnectMqtt() {
+    if (deviceTarget != null) {
+      captureExceptions("closing mqtt publisher", deviceTarget::close);
+      captureExceptions("shutting down mqtt publisher executor", deviceTarget::shutdown);
+      deviceTarget = null;
+    }
+  }
+
   @Override
   public void initializeMqtt() {
     checkNotNull(config.deviceId, "configuration deviceId not defined");
@@ -664,6 +742,24 @@ public class Pubber extends ManagerBase implements PubberHost {
     return ofNullable(catchToNull(() -> metadata.gateway.gateway_id)).orElse(deviceId);
   }
 
+  private void registerMessageHandlers() {
+    deviceTarget.registerHandler(CONFIG_TOPIC, this::configHandler, Config.class);
+    String gatewayId = PubberHost.getGatewayId(deviceId, config);
+    if (isGatewayDevice) {
+      // In this case, this is the gateway so register the appropriate error handler directly.
+      deviceTarget.registerHandler(ERRORS_TOPIC, this::errorHandler, GatewayError.class);
+    } else if (gatewayId != null) {
+      // In this case, this is a proxy device with a gateway, so register handlers accordingly.
+      MqttDevice gatewayTarget = new MqttDevice(gatewayId, deviceTarget);
+      gatewayTarget.registerHandler(CONFIG_TOPIC, this::gatewayHandler, Config.class);
+      gatewayTarget.registerHandler(ERRORS_TOPIC, this::errorHandler, GatewayError.class);
+    }
+  }
+
+  public MqttDevice getMqttDevice(String proxyId) {
+    return new MqttDevice(proxyId, deviceTarget);
+  }
+
   @Override
   public byte[] ensureKeyBytes() {
     if (config.keyBytes == null) {
@@ -675,6 +771,21 @@ public class Pubber extends ManagerBase implements PubberHost {
     return (byte[]) config.keyBytes;
   }
 
+  private void connect() {
+    try {
+      warn("Creating new config latch for " + deviceId);
+      configLatch = new CountDownLatch(1);
+      deviceTarget.connect();
+      info("Connection complete.");
+      workingEndpoint = toJsonString(config.endpoint);
+    } catch (Exception e) {
+      throw new RuntimeException("Connection error", e);
+    }
+  }
+
+  public void publisherConfigLog(String phase, Exception e, String targetId) {
+    publisherHandler("config", phase, e, targetId);
+  }
 
   @Override
   public void publisherException(Exception toReport) {
@@ -694,6 +805,228 @@ public class Pubber extends ManagerBase implements PubberHost {
     }
   }
 
+  private void publisherHandler(String type, String phase, Throwable cause, String targetId) {
+    if (cause != null) {
+      error("Error receiving message " + type, cause);
+      if (isTrue(config.options.barfConfig)) {
+        error("Restarting system because of restart-on-error configuration setting");
+        deviceManager.systemLifecycle(SystemMode.RESTART);
+      }
+    }
+    String usePhase = isTrue(options.badCategory) ? "apply" : phase;
+    String category = format(SYSTEM_CATEGORY_FORMAT, type, usePhase);
+    Entry report = entryFromException(category, cause);
+    deviceManager.localLog(report);
+    publishLogMessage(report, targetId);
+    ifTrueThen(deviceId.equals(targetId), () -> registerSystemStatus(report));
+  }
+
+  private void registerSystemStatus(Entry report) {
+    deviceState.system.status = report;
+    markStateDirty();
+  }
+
+  /**
+   * Issue a state update in response to a received config message. This will optionally add a
+   * synthetic delay in so that testing infrastructure can test that related sequence tests handle
+   * this case appropriately.
+   */
+  private void publishConfigStateUpdate() {
+    if (isTrue(config.options.configStateDelay)) {
+      delayNextStateUpdate();
+    }
+    publishAsynchronousState();
+  }
+
+  private void delayNextStateUpdate() {
+    // Calculate a synthetic last state time that factors in the optional delay.
+    long syntheticType = System.currentTimeMillis() - STATE_THROTTLE_MS + FORCED_STATE_TIME_MS;
+    // And use the synthetic time iff it's later than the actual last state time.
+    lastStateTimeMs = Math.max(lastStateTimeMs, syntheticType);
+  }
+
+  private Entry entryFromException(String category, Throwable e) {
+    boolean success = e == null;
+    Entry entry = new Entry();
+    entry.category = category;
+    entry.timestamp = getNow();
+    entry.message = success ? "success"
+        : e.getMessage() != null ? e.getMessage() : e.getClass().getSimpleName();
+    entry.detail = success ? null : exceptionDetail(e);
+    Level successLevel = Category.LEVEL.computeIfAbsent(category, key -> Level.INFO);
+    entry.level = (success ? successLevel : Level.ERROR).value();
+    return entry;
+  }
+
+  private String exceptionDetail(Throwable e) {
+    StringBuilder buffer = new StringBuilder();
+    while (e != null) {
+      buffer.append(e).append(';');
+      e = e.getCause();
+    }
+    return buffer.toString();
+  }
+
+  private void configHandler(Config config) {
+    try {
+      configPreprocess(deviceId, config);
+      debug(format("Config update %s%s", deviceId, deviceManager.getTestingTag()),
+          toJsonString(config));
+      processConfigUpdate(config);
+      if (configLatch.getCount() > 0) {
+        warn("Received config for config latch " + deviceId);
+        configLatch.countDown();
+      }
+      publisherConfigLog("apply", null, deviceId);
+    } catch (Exception e) {
+      publisherConfigLog("apply", e, deviceId);
+    }
+    publishConfigStateUpdate();
+  }
+
+  private void gatewayHandler(Config gatewayConfig) {
+    warn("Ignoring configuration for gateway " + PubberHost.getGatewayId(deviceId, config));
+  }
+
+  private void errorHandler(GatewayError error) {
+    warn(format("%s for %s: %s", error.error_type, error.device_id, error.description));
+  }
+
+  void configPreprocess(String targetId, Config configMsg) {
+    String gatewayId = PubberHost.getGatewayId(targetId, config);
+    String suffix = ifNotNullGet(gatewayId, x -> "_" + targetId, "");
+    String deviceType = ifNotNullGet(gatewayId, x -> "Proxy", "Device");
+    info(format("%s %s config handler", deviceType, targetId));
+    File configOut = new File(outDir, format("%s.json", traceTimestamp("config" + suffix)));
+    toJsonFile(configOut, configMsg);
+  }
+
+  private void processConfigUpdate(Config configMsg) {
+    try {
+      // Grab this to make state-after-config updates monolithic.
+      stateLock.lock();
+    } catch (Exception e) {
+      throw new RuntimeException("While acquiring state lock", e);
+    }
+
+    try {
+      if (configMsg != null) {
+        if (configMsg.system == null && isTrue(config.options.barfConfig)) {
+          error("Empty config system block and configured to restart on bad config!");
+          deviceManager.systemLifecycle(SystemMode.RESTART);
+        }
+        GeneralUtils.copyFields(configMsg, deviceConfig, true);
+        info(format("%s received config %s", getTimestamp(), isoConvert(configMsg.timestamp)));
+        deviceManager.updateConfig(configMsg);
+        extractEndpointBlobConfig();
+      } else {
+        info(getTimestamp() + " defaulting empty config");
+      }
+      updateInterval(DEFAULT_REPORT_SEC);
+    } finally {
+      stateLock.unlock();
+    }
+  }
+
+  // TODO: Consider refactoring this to either return or change an instance variable, not both.
+  EndpointConfiguration extractEndpointBlobConfig() {
+    extractedEndpoint = null;
+    if (deviceConfig.blobset == null) {
+      return null;
+    }
+    try {
+      String iotConfig = extractConfigBlob(IOT_ENDPOINT_CONFIG.value());
+      extractedEndpoint = fromJsonString(iotConfig, EndpointConfiguration.class);
+      if (extractedEndpoint != null) {
+        if (deviceConfig.blobset.blobs.containsKey(IOT_ENDPOINT_CONFIG.value())) {
+          BlobBlobsetConfig config = deviceConfig.blobset.blobs.get(IOT_ENDPOINT_CONFIG.value());
+          extractedEndpoint.generation = config.generation;
+        }
+      }
+    } catch (Exception e) {
+      throw new RuntimeException("While extracting endpoint blob config", e);
+    }
+    return extractedEndpoint;
+  }
+
+  private void removeBlobsetBlobState(SystemBlobsets blobId) {
+    if (deviceState.blobset == null) {
+      return;
+    }
+
+    if (deviceState.blobset.blobs.remove(blobId.value()) == null) {
+      return;
+    }
+
+    if (deviceState.blobset.blobs.isEmpty()) {
+      deviceState.blobset = null;
+    }
+
+    markStateDirty();
+  }
+
+  void maybeRedirectEndpoint() {
+    String redirectRegistry = config.options.redirectRegistry;
+    String currentSignature = toJsonString(config.endpoint);
+    String extractedSignature =
+        redirectRegistry == null ? toJsonString(extractedEndpoint)
+            : redirectedEndpoint(redirectRegistry);
+
+    if (extractedSignature == null) {
+      attemptedEndpoint = null;
+      removeBlobsetBlobState(IOT_ENDPOINT_CONFIG);
+      return;
+    }
+
+    BlobBlobsetState endpointState = ensureBlobsetState(IOT_ENDPOINT_CONFIG);
+
+    if (extractedSignature.equals(currentSignature)
+        || extractedSignature.equals(attemptedEndpoint)) {
+      return; // No need to redirect anything!
+    }
+
+    if (extractedEndpoint != null) {
+      if (!Objects.equals(endpointState.generation, extractedEndpoint.generation)) {
+        notice("Starting new endpoint generation");
+        endpointState.phase = null;
+        endpointState.status = null;
+        endpointState.generation = extractedEndpoint.generation;
+      }
+
+      if (extractedEndpoint.error != null) {
+        attemptedEndpoint = extractedSignature;
+        endpointState.phase = BlobPhase.FINAL;
+        Exception applyError = new RuntimeException(extractedEndpoint.error);
+        endpointState.status = exceptionStatus(applyError, Category.BLOBSET_BLOB_APPLY);
+        publishSynchronousState();
+        return;
+      }
+    }
+
+    info("New config blob endpoint detected:\n" + stringify(parseJson(extractedSignature)));
+
+    try {
+      attemptedEndpoint = extractedSignature;
+      endpointState.phase = BlobPhase.APPLY;
+      publishSynchronousState();
+      resetConnection(extractedSignature);
+      persistEndpoint(extractedEndpoint);
+      endpointState.phase = BlobPhase.FINAL;
+      markStateDirty();
+    } catch (Exception e) {
+      try {
+        error("Reconfigure failed, attempting connection to last working endpoint", e);
+        endpointState.phase = BlobPhase.FINAL;
+        endpointState.status = exceptionStatus(e, Category.BLOBSET_BLOB_APPLY);
+        resetConnection(workingEndpoint);
+        publishAsynchronousState();
+        notice("Endpoint connection restored to last working endpoint");
+      } catch (Exception e2) {
+        throw new RuntimeException("While restoring working endpoint", e2);
+      }
+      error("While redirecting connection endpoint", e);
+    }
+  }
 
   @Override
   public void persistEndpoint(EndpointConfiguration endpoint) {
@@ -713,16 +1046,6 @@ public class Pubber extends ManagerBase implements PubberHost {
   }
 
   @Override
-  public AtomicBoolean getStateDirty() {
-    return null;
-  }
-
-  @Override
-  public SchemaVersion getTargetSchema() {
-    return null;
-  }
-
-  @Override
   public void resetConnection(String targetEndpoint) {
     try {
       config.endpoint = fromJsonString(targetEndpoint,
@@ -734,6 +1057,189 @@ public class Pubber extends ManagerBase implements PubberHost {
     } catch (Exception e) {
       throw new RuntimeException("While resetting connection", e);
     }
+  }
+
+  private Entry exceptionStatus(Exception e, String category) {
+    Entry entry = new Entry();
+    entry.message = e.getMessage();
+    entry.detail = stackTraceString(e);
+    entry.category = category;
+    entry.level = Level.ERROR.value();
+    entry.timestamp = getNow();
+    return entry;
+  }
+
+  private BlobBlobsetState ensureBlobsetState(SystemBlobsets iotEndpointConfig) {
+    deviceState.blobset = ofNullable(deviceState.blobset).orElseGet(BlobsetState::new);
+    deviceState.blobset.blobs = ofNullable(deviceState.blobset.blobs).orElseGet(HashMap::new);
+    return deviceState.blobset.blobs.computeIfAbsent(iotEndpointConfig.value(),
+        key -> new BlobBlobsetState());
+  }
+
+  private String getClientId(String forRegistry) {
+    String cloudRegion = SiteModel.parseClientId(config.endpoint.client_id).cloudRegion;
+    return SiteModel.getClientId(config.iotProject, cloudRegion, forRegistry, deviceId);
+  }
+
+  private String extractConfigBlob(String blobName) {
+    // TODO: Refactor to get any blob meta parameters.
+    try {
+      if (deviceConfig == null || deviceConfig.blobset == null
+          || deviceConfig.blobset.blobs == null) {
+        return null;
+      }
+      BlobBlobsetConfig blobBlobsetConfig = deviceConfig.blobset.blobs.get(blobName);
+      if (blobBlobsetConfig != null && BlobPhase.FINAL.equals(blobBlobsetConfig.phase)) {
+        return PubberHost.acquireBlobData(blobBlobsetConfig.url, blobBlobsetConfig.sha256);
+      }
+      return null;
+    } catch (Exception e) {
+      EndpointConfiguration endpointConfiguration = new EndpointConfiguration();
+      endpointConfiguration.error = e.toString();
+      return stringify(endpointConfiguration);
+    }
+  }
+
+  private void publishLogMessage(Entry logEntry, String targetId) {
+    deviceManager.publishLogMessage(logEntry, targetId);
+  }
+
+  private void publishAsynchronousState() {
+    if (stateLock.tryLock()) {
+      try {
+        long soonestAllowedStateUpdate = lastStateTimeMs + STATE_THROTTLE_MS;
+        long delay = soonestAllowedStateUpdate - System.currentTimeMillis();
+        debug(format("State update defer %dms", delay));
+        if (delay > 0) {
+          markStateDirty(delay);
+        } else {
+          publishStateMessage();
+        }
+      } finally {
+        stateLock.unlock();
+      }
+    } else {
+      markStateDirty(-1);
+    }
+  }
+
+  void publishSynchronousState() {
+    try {
+      stateLock.lock();
+      publishStateMessage();
+    } catch (Exception e) {
+      throw new RuntimeException("While sending synchronous state", e);
+    } finally {
+      stateLock.unlock();
+    }
+  }
+
+  boolean publisherActive() {
+    return deviceTarget != null && deviceTarget.isActive();
+  }
+
+  private void publishStateMessage() {
+    if (!publisherActive()) {
+      markStateDirty(-1);
+      return;
+    }
+    stateDirty.set(false);
+    deviceState.timestamp = getNow();
+    info(format("Update state %s last_config %s", isoConvert(deviceState.timestamp),
+        isoConvert(deviceState.system.last_config)));
+    publishStateMessage(isTrue(options.badState) ? deviceState.system : deviceState);
+  }
+
+  private void publishStateMessage(Object stateToSend) {
+    try {
+      stateLock.lock();
+      publishStateMessageRaw(stateToSend);
+    } finally {
+      stateLock.unlock();
+    }
+  }
+
+  private void publishStateMessageRaw(Object stateToSend) {
+    if (configLatch == null || configLatch.getCount() > 0) {
+      warn("Dropping state update until config received...");
+      return;
+    }
+
+    long delay = lastStateTimeMs + STATE_THROTTLE_MS - System.currentTimeMillis();
+    if (delay > 0) {
+      warn(format("State update delay %dms", delay));
+      safeSleep(delay);
+    }
+
+    lastStateTimeMs = System.currentTimeMillis();
+    CountDownLatch latch = new CountDownLatch(1);
+
+    try {
+      debug(format("State update %s%s", deviceId, deviceManager.getTestingTag()),
+          toJsonString(stateToSend));
+    } catch (Exception e) {
+      throw new RuntimeException("While converting new device state", e);
+    }
+
+    publishDeviceMessage(deviceId, stateToSend, () -> {
+      lastStateTimeMs = System.currentTimeMillis();
+      latch.countDown();
+    });
+    try {
+      if (shouldSendState() && !latch.await(WAIT_TIME_SEC, TimeUnit.SECONDS)) {
+        throw new RuntimeException("Timeout waiting for state send");
+      }
+    } catch (Exception e) {
+      throw new RuntimeException(format("While waiting for %s state send latch", deviceId), e);
+    }
+  }
+
+  private boolean shouldSendState() {
+    return !isGetTrue(() -> config.options.noState);
+  }
+
+  private void publishDeviceMessage(Object message) {
+    publishDeviceMessage(deviceId, message);
+  }
+
+  private void publishDeviceMessage(String targetId, Object message) {
+    publishDeviceMessage(targetId, message, null);
+  }
+
+  private void publishDeviceMessage(String targetId, Object message, Runnable callback) {
+    String topicSuffix = MESSAGE_TOPIC_SUFFIX_MAP.get(message.getClass());
+    if (topicSuffix == null) {
+      error("Unknown message class " + message.getClass());
+      return;
+    }
+
+    if (!shouldSendState() && topicSuffix.equals(STATE_TOPIC)) {
+      warn("Squelching state update as per configuration");
+      return;
+    }
+
+    if (deviceTarget == null) {
+      error("publisher not active");
+      return;
+    }
+
+    PubberHost.augmentDeviceMessage(message, getNow(), isTrue(options.badVersion));
+    Object downgraded = downgradeMessage(message);
+    deviceTarget.publish(targetId, topicSuffix, downgraded, callback);
+    String messageBase = topicSuffix.replace("/", "_");
+    String gatewayId = PubberHost.getGatewayId(targetId, config);
+    String suffix = ifNotNullGet(gatewayId, x -> "_" + targetId, "");
+    File messageOut = new File(outDir, format("%s.json", traceTimestamp(messageBase + suffix)));
+    try {
+      toJsonFile(messageOut, downgraded);
+    } catch (Exception e) {
+      throw new RuntimeException("While writing " + messageOut.getAbsolutePath(), e);
+    }
+  }
+
+  private Object downgradeMessage(Object message) {
+    MessageDowngrader messageDowngrader = new MessageDowngrader(SubType.STATE.value(), message);
+    return ifNotNullGet(targetSchema, messageDowngrader::downgrade, message);
   }
 
   @Override
@@ -767,12 +1273,16 @@ public class Pubber extends ManagerBase implements PubberHost {
     cloudLog(message, Level.DEBUG);
   }
 
+  private void debug(String message, String detail) {
+    cloudLog(message, Level.DEBUG, detail);
+  }
+
   @Override
   public void info(String message) {
     cloudLog(message, Level.INFO);
   }
 
-  public void notice(String message) {
+  private void notice(String message) {
     cloudLog(message, Level.NOTICE);
   }
 
@@ -797,93 +1307,4 @@ public class Pubber extends ManagerBase implements PubberHost {
     deviceManager.localLog(message, Level.TRACE, getTimestamp(), stackTraceString(e));
   }
 
-  @Override
-  public void setLastStateTimeMs(long lastStateTimeMs) {
-
-  }
-
-  @Override
-  public long getLastStateTimeMs() {
-    return 0;
-  }
-
-  @Override
-  public CountDownLatch getConfigLatch() {
-    return null;
-  }
-
-  @Override
-  public File getOutDir() {
-    return null;
-  }
-
-  @Override
-  public Lock getStateLock() {
-    return null;
-  }
-
-  @Override
-  public EndpointConfiguration getExtractedEndpoint() {
-    return null;
-  }
-
-  @Override
-  public void setExtractedEndpoint(EndpointConfiguration endpointConfiguration) {
-
-  }
-
-  @Override
-  public String getWorkingEndpoint() {
-    return "";
-  }
-
-  @Override
-  public void setAttemptedEndpoint(String s) {
-
-  }
-
-  @Override
-  public String getAttemptedEndpoint() {
-    return "";
-  }
-
-  @Override
-  public State getDeviceState() {
-    return deviceState;
-  }
-
-  @Override
-  public Config getDeviceConfig() {
-    return deviceConfig;
-  }
-
-  @Override
-  public DeviceManager getDeviceManager() {
-    return deviceManager;
-  }
-
-  @Override
-  public MqttDevice getDeviceTarget() {
-    return deviceTarget;
-  }
-
-  @Override
-  public void setDeviceTarget(MqttDevice deviceTarget) {
-    this.deviceTarget = deviceTarget;
-  }
-
-  @Override
-  public boolean isGatewayDevice() {
-    return isGatewayDevice;
-  }
-
-  @Override
-  public void setWorkingEndpoint(String jsonString) {
-    this.workingEndpoint = jsonString;
-  }
-
-  @Override
-  public void setConfigLatch(CountDownLatch countDownLatch) {
-    this.configLatch = countDownLatch;
-  }
 }
